+++ keyboards/helix/helix.h
@@ -18,5 +18,5 @@
     #include "pico.h"
 #endif
 #ifdef KEYBOARD_helix_rev2_latam
-    #include "rev2.h"
+    #include "rev2_latam.h"
 #endif
