+++ keyboards/helix/rev2_latam/back/rules.mk
@@ -1,2 +1,1 @@
 LED_BACK_ENABLE = yes
-SPLIT_KEYBOARD = yes
