--- keyboards/helix/rev2_latam/keymaps/edvorakjp/keymap_4rows.c
@@ -1,41 +1,0 @@
-#include "helix.h"
-#include "keymap_xrows.h"
-
-const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
-
-  [_EDVORAK] = LAYOUT_kc(
-  //,----+----+-----+-----+-----+----.           ,----+-----+-----+-----+----+----.
-     GRV ,QUOT,COMM , DOT ,  Y  , Q  ,             F  ,  G  ,  R  ,  W  , P  ,BSLS,
-  //|----+----+-----+-----+-----+----|           |----+-----+-----+-----+----+----|
-     EQL , A  ,LA(O),LG(E),LC(I), U  ,             D  ,RS(T),RG(N),RA(S), M  ,MINS,
-  //|----+----+-----+-----+-----+----|           |----+-----+-----+-----+----+----|
-     ESC ,SCLN,  X  ,  C  ,  V  , Z  ,             H  ,  J  ,  K  ,  L  , B  ,SLSH,
-  //|----+----+-----+-----+-----+----+----. ,----+----+-----+-----+-----+----+----|
-     HOME,PGDN,PGUP , END ,TMB1 ,TMB2,TMB3,  TMB4,TMB5,TMB6 ,LEFT ,DOWN , UP ,RGHT
-  //`----+----+-----+-----+-----+----+----/ \----+----+-----+-----+-----+----+----'
-  ),
-
-  [_LOWER] = LAYOUT_kc(
-  //,----+----+------+------+------+----.           ,----+------+------+-------+----+----.
-         ,    , LCBR , LBRC , LPRN ,    ,                , RPRN , RBRC , RCBR  ,    ,    ,
-  //|----+----+------+------+------+----|           |----+------+------+-------+----+----|
-      F1 , F2 ,LA(F3),LG(F4),LC(F5), F6 ,             F7 ,RS(F8),RG(F9),RA(F10),F11 ,F12 ,
-  //|----+----+------+------+------+----|           |----+------+------+-------+----+----|
-     PSCR,    , HOME , PGDN , PGUP ,END ,            LEFT, DOWN ,  UP  , RGHT  ,    ,    ,
-  //|----+----+------+------+------+----+----. ,----+----+------+------+-------+----+----|
-         ,    ,      ,      ,      ,    , NO ,  MAC ,TMB7,      ,      ,       ,    ,
-  //`----+----+------+------+------+----+----/ \----+----+------+------+-------+----+----'
-  ),
-
-  [_RAISE] = LAYOUT_kc(
-  //,----+----+-----+-----+-----+----.           ,----+-----+-----+-----+----+----.
-         ,    ,EXLM , AT  ,HASH ,DLR ,            PERC,CIRC ,AMPR ,ASTR ,    ,    ,
-  //|----+----+-----+-----+-----+----|           |----+-----+-----+-----+----+----|
-         , 1  ,LA(2),LG(3),LC(4), 5  ,             6  ,RS(7),RG(8),RA(9), 0  ,    ,
-  //|----+----+-----+-----+-----+----|           |----+-----+-----+-----+----+----|
-     PSCR,    ,HOME ,PGDN ,PGUP ,END ,            LEFT,DOWN , UP  ,RGHT ,    ,    ,
-  //|----+----+-----+-----+-----+----+----. ,----+----+-----+-----+-----+----+----|
-         ,    ,     ,     ,     ,    ,WIN ,   NO ,    ,     ,     ,     ,    ,
-  //`----+----+-----+-----+-----+----+----/ \----+----+-----+-----+-----+----+----'
-  )
-};
