--- keyboards/helix/rev2_latam/keymaps/edvorakjp/keymap_5rows.c
@@ -1,47 +1,0 @@
-#include "helix.h"
-#include "keymap_xrows.h"
-
-const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
-
-  [_EDVORAK] = LAYOUT_kc(
-  //,----+----+-----+-----+-----+----.           ,----+-----+-----+-----+----+----.
-     ESC ,EXLM, AT  ,HASH , DLR ,PERC,            CIRC,AMPR ,ASTR ,LPRN ,RPRN,BSPC,
-  //|----+----+-----+-----+-----+----|           |----+-----+-----+-----+----+----|
-     GRV ,QUOT,COMM , DOT ,  Y  , Q  ,             F  ,  G  ,  R  ,  W  , P  ,BSLS,
-  //|----+----+-----+-----+-----+----|           |----+-----+-----+-----+----+----|
-     EQL , A  ,LA(O),LG(E),LC(I), U  ,             D  ,RS(T),RG(N),RA(S), M  ,MINS,
-  //|----+----+-----+-----+-----+----+----. ,----+----+-----+-----+-----+----+----|
-     TAB ,SCLN,  X  ,  C  ,  V  , Z  , NO ,   NO , H  ,  J  ,  K  ,  L  , B  ,SLSH,
-  //|----+----+-----+-----+-----+----+----| |----+----+-----+-----+-----+----+----|
-     HOME,PGDN,PGUP , END ,TMB1 ,TMB2,TMB3,  TMB4,TMB5,TMB6 ,LEFT ,DOWN , UP ,RGHT
-  //`----+----+-----+-----+-----+----+----/ \----+----+-----+-----+-----+----+----'
-  ),
-
-  [_LOWER] = LAYOUT_kc(
-  //,----+----+------+------+------+----.           ,----+------+------+-------+----+----.
-     PSCR,    ,      ,      ,      ,    ,                ,      ,      ,       ,    ,    ,
-  //|----+----+------+------+------+----|           |----+------+------+-------+----+----|
-         ,    , LCBR , LBRC , LPRN ,    ,                , RPRN , RBRC , RCBR  ,    ,    ,
-  //|----+----+------+------+------+----|           |----+------+------+-------+----+----|
-      F1 , F2 ,LA(F3),LG(F4),LC(F5), F6 ,             F7 ,RS(F8),RG(F9),RA(F10),F11 ,F12 ,
-  //|----+----+------+------+------+----+----. ,----+----+------+------+-------+----+----|
-         ,    , HOME , PGDN , PGUP ,END ,    ,      ,LEFT, DOWN ,  UP  , RGHT  ,    ,    ,
-  //|----+----+------+------+------+----+----| |----+----+------+------+-------+----+----|
-         ,    ,      ,      ,      ,    , NO ,  MAC ,TMB7,      ,      ,       ,    ,
-  //`----+----+------+------+------+----+----/ \----+----+------+------+-------+----+----'
-  ),
-
-  [_RAISE] = LAYOUT_kc(
-  //,----+----+-----+-----+-----+----.           ,----+-----+-----+-----+----+----.
-     PSCR,    ,     ,     ,     ,    ,                ,     ,     ,     ,    ,    ,
-  //|----+----+-----+-----+-----+----|           |----+-----+-----+-----+----+----|
-         ,    ,EXLM , AT  ,HASH ,DLR ,            PERC,CIRC ,AMPR ,ASTR ,    ,    ,
-  //|----+----+-----+-----+-----+----|           |----+-----+-----+-----+----+----|
-         , 1  ,LA(2),LG(3),LC(4), 5  ,             6  ,RS(7),RG(8),RA(9), 0  ,    ,
-  //|----+----+-----+-----+-----+----+----. ,----+----+-----+-----+-----+----+----|
-         ,    ,HOME ,PGDN ,PGUP ,END ,    ,      ,LEFT,DOWN , UP  ,RGHT ,    ,    ,
-  //|----+----+-----+-----+-----+----+----| |----+----+-----+-----+-----+----+----|
-         ,    ,     ,     ,     ,    ,WIN ,   NO ,    ,     ,     ,     ,    ,
-  //`----+----+-----+-----+-----+----+----/ \----+----+-----+-----+-----+----+----'
-  )
-};
