--- keyboards/helix/rev2_latam/keymaps/froggy/readme.md
@@ -1,83 +1,0 @@
-Froggy -one hand Helix-
-======
-
-![Imgur](https://i.imgur.com/S1Dw3XW.jpg)
-
-## Features
-It is a one-handed keyboard with reference to Frogpad. Layout Designed by タクマ ([@humid](https://twitter.com/humid)).
-
-## Layout
-### Base
-```
-,-----------------------------------------.
-|  C+z |   ;  |   [  |   (  |   <  |   {  |
-|------+------+------+------+------+------|
-| KANA |   P  |   K  |   R  |   A  |   F  |
-|------+------+------+------+------+------|
-|  BS  |   D  |   T  |   H  |   E  |   O  |
-|------+------+------+------+------+------+------.
-| Shift|   Y  |   S  |   N  |   I  |   U  | Space|
-|------+------+------+------+------+------+------|
-| Ctrl |  Alt |  win |  Sym |  Num |  OPT |  Ent |
-`------------------------------------------------'
-```
-
-### Opt
-```
-,-----------------------------------------.
-|  Esc |   :  |   ]  |   )  |   >  |   }  |
-|------+------+------+------+------+------|
-| EISU |   J  |   M  |   B  |   '  |  Tab |
-|------+------+------+------+------+------|
-|   .  |   V  |   C  |   L  |   Z  |   Q  |
-|------+------+------+------+------+------+------.
-|      |   X  |   G  |   W  |   -  |  Del |  Esc |
-|------+------+------+------+------+------+------|
-|      |      |      |   ,  | DTOP |      |      |
-`------------------------------------------------'
-```
-
-### Num
-```
-,-----------------------------------------.
-|      |      | Func | home |  End |      |
-|------+------+------+------+------+------|
-|      |   *  |   7  |   8  |   9  |   -  |
-|------+------+------+------+------+------|
-|   .  |   /  |   4  |   5  |   6  |   +  |
-|------+------+------+------+------+------+------.
-|  LN  |   0  |   1  |   2  |   3  |C+S+F1|      |
-|------+------+------+------+------+------+------|
-|      |      |      |   ,  |      |      |      |
-`------------------------------------------------'
-```
-
-### Sym
-```
-,-----------------------------------------.
-|  Ins |  GRV |      |  PU  |  PD  |   ^  |
-|------+------+------+------+------+------|
-|      |   \  |   #  |   =  |   ?  |   %  |
-|------+------+------+------+------+------|
-|      |   $  |  upA |   @  |   !  |   |  |
-|------+------+------+------+------+------+------.
-|  CL  |  <-  |  dwA |  ->  |   _  |   &  |      |
-|------+------+------+------+------+------+------|
-|      |      |  PS  |      |   ~  |      |      |
- `-----------------------------------------------'
-```
-
-### Func
-```
-,-----------------------------------------.
-|RGBRST|  Hue |      |  RST |  Mac |  Win |
-|------+------+------+------+------+------|
-| RGB1 | VAL+ |  F7  |  F8  |  F9  |      |
-|------+------+------+------+------+------|
-| RGB2 | VAL- |  F4  |  F5  |  F6  |  F12 |
-|------+------+------+------+------+------+------.
-| RGB3 |  F10 |  F1  |  F2  |  F3  |  F11 |      |
-|------+------+------+------+------+------+------|
-|RGBOFF|      |      |      |      |      |      |
-`------------------------------------------------'
-```
