+++ keyboards/helix/rev2_latam/keymaps/latam/rules.mk
@@ -6,19 +6,18 @@
 #   See TOP/docs/config_options.md for more information.
 #
 LTO_ENABLE = no  # if firmware size over limit, try this option
-
+EXTRAKEY_ENABLE = yes    # Audio control and System control
+# AUDIO_ENABLE = yes       # Audio output on port C6
 # Helix Spacific Build Options
 # you can uncomment and edit follows 7 Variables
 #  jp: 以下の7つの変数を必要に応じて編集し、コメントアウトをはずします。
 # HELIX_ROWS = 5              # Helix Rows is 4 or 5
-OLED_ENABLE = yes            # OLED_ENABLE
+# OLED_ENABLE = no            # OLED_ENABLE
 # LOCAL_GLCDFONT = no         # use each keymaps "helixfont.h" insted of "common/glcdfont.c"
-LED_BACK_ENABLE = yes        # LED backlight (Enable WS2812 RGB underlight.)
+# LED_BACK_ENABLE = no        # LED backlight (Enable WS2812 RGB underlight.)
 # LED_UNDERGLOW_ENABLE = no   # LED underglow (Enable WS2812 RGB underlight.)
-LED_ANIMATIONS = yes        # LED animations
+# LED_ANIMATIONS = yes        # LED animations
 # IOS_DEVICE_ENABLE = no      # connect to IOS device (iPad,iPhone)
-
-SRC += led_test_init.c
 
 # convert Helix-specific options (that represent combinations of standard options)
 #   into QMK standard options.
