--- keyboards/helix/rev2_latam/keymaps/led_test/keymap.c
@@ -1,1 +1,0 @@
-#include "../default/keymap.c"
