--- keyboards/helix/rev2_latam/keymaps/led_test/led_test_init.c
@@ -1,6 +1,0 @@
-#include QMK_KEYBOARD_H
-
-void keyboard_post_init_user(void) {
-    rgblight_enable_noeeprom();
-    rgblight_mode_noeeprom(RGBLIGHT_MODE_RGB_TEST);
-}
