--- keyboards/helix/rev2_latam/keymaps/xulkal/rules.mk
@@ -1,13 +1,0 @@
-RGBLIGHT_ENABLE = yes
-# Enable RGBLIGHT Animations
-OPT_DEFS += -DRGBLIGHT_ANIMATIONS
-# Helix specific define for correct RGBLED_NUM
-OPT_DEFS += -DRGBLED_BACK
-
-OLED_DRIVER_ENABLE = yes
-# Helix specific font file
-OPT_DEFS += -DOLED_FONT_H=\"common/glcdfont.c\"
-# Xulkal specific oled define
-OPT_DEFS += -DOLED_90ROTATION
-
-SPLIT_KEYBOARD = yes
